vite installして、errorが消えるところまで
