+++ examples/playground/vite.config.js
@@ -2,6 +2,8 @@ import path from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { defineConfig } from 'vite'
 
+import chibivue from '../../packages/@extensions/vite-plugin-chibivue'
+
 const dirname = path.dirname(fileURLToPath(new URL(import.meta.url)))
 export default defineConfig({
   resolve: {
@@ -9,4 +11,5 @@ export default defineConfig({
       chibivue: path.resolve(dirname, '../../packages'),
     },
   },
+  plugins: [chibivue()],
 })
