+++ packages/@extensions/vite-plugin-chibivue/index.ts
@@ -1,0 +1,14 @@
+import { createFilter, type Plugin } from 'vite'
+
+export default function vitePluginChibivue(): Plugin {
+  const filter = createFilter(/\.vue$/)
+
+  return {
+    name: 'vite:chibivue',
+
+    transform(code, id) {
+      if (!filter(id)) return
+      return { code: `export default {}` }
+    },
+  }
+}
